fix(publish): move verify endpoint to /verify subpath per spec
- POST handler was at [domainId]/route.ts (wrong URL: /domains/{id})
- Spec requires POST /api/drives/{driveId}/domains/{domainId}/verify
- Move POST + tests to [domainId]/verify/route.ts + verify/__tests__/
- Restore [domainId]/route.ts to DELETE-only (matches PR1 state)
- Update UI handleVerifyDomain to call /domains/{id}/verify
